Extract TestDelwaqCoupling logic to separate template
